requirements.txt
flask
werkzeug
torch
torchvision
opencv-python
numpy
pillow